refactor: main-login page 코드리뷰 합치기
